update idl to support meta and doc (#23)
idl comments start with `//` and are always single line comments. The lines are combined to the description attribute. A `xxDoc` filter will format the doc according to the selected language.
The meta information start with `@` and the text need to follow the YAML spec. Several lines are combined to a YAML object. The YAML information is available in the template as map of strings.