precision types (#20)
* wip: add int32, int64, float32, float64 support

- model
- schema
- idl
- go filters

* add cpp precision types

* py support for precision types

* qt support for precision types

* ts support for precision type

* ue support for precision types

* go filter support precision types

* update idl to support meta and doc (#23)

idl comments start with `//` and are always single line comments. The lines are combined to the description attribute. A `xxDoc` filter will format the doc according to the selected language.
The meta information start with `@` and the text need to follow the YAML spec. Several lines are combined to a YAML object. The YAML information is available in the template as map of strings.

* wip: add int32, int64, float32, float64 support

- model
- schema
- idl
- go filters

* update antlr deps

* update precision tests

* Update go.sum