support for system.meta in solution/layer doc
We want to push meta information from a solution doc into the template. This will be available on the system (`$.System.Meta`). The meta information on the doc layer will be merged with the meta from the layer element.
Additional the cli shall later be able to add meta on the cmd line. (`--meta`)