feat(python): add filter for test value
adds a filter which, for simple types sets some value that is different
than the default value.
For enum it returns te second value (if present).
For the arrays it returns an element of that type
For structs it returns the default value.
This is used by a python template for #35 ticket