feat(qt): add test value filter for qt
In Qt template we are missing some tests for our generated code.
With this filter, setting a property (signal value) different that
default can be generated. It allows us to test our generated interfaces
for which we use goldenmaster - the apis that cover hopefully most of
the uscases. Also provides generated tests for the users of ApiGear.
This filter is required to solve the #90 ticket #89 in template-qtcpp.