filters(jni): add filters for java type signature
When finding java classes fields or methods in cpp the type has specific
signature. This filter returns proper signature type, including the full
path of generated modules containing custom types.
The apigear generated java code structure is reflected in the filter,
have in mind that if that changes, the filter needs to be adjusted.