filters(jni): env name type
Java env contains some functions that have variants for different data
type. With this filter you can get proper this type part of name
function that you plan to use. It couldn' be eaisly converted from the
type, or achived with other filters.
It skips the Array part - it needs to be added manulally, as for some
functions it is followed by "region" word and for some not.